[Refactor] #94 : Separate ItemIdentifierType of datasource into SubCategory and StoreListItem using enum associated value
